+++ examples/ff-pm/main_ff.cc
@@ -24,161 +24,118 @@
 #include <config.h>
 
 #include <ctime>
-#include <fstream>
 #include <iostream>
-#include <string>
-
-bool printstuff = false;
 
 #include <dune/common/parallel/mpihelper.hh>
 #include <dune/common/timer.hh>
-#include <dune/grid/io/file/dgfparser/dgfexception.hh>
-#include <dune/grid/io/file/vtk.hh>
 #include <dune/istl/io.hh>
 
-#include <dumux/common/defaultusagemessage.hh>
+#include <dumux/assembly/diffmethod.hh>
+#include <dumux/multidomain/fvassembler.hh>
+#include <dumux/multidomain/traits.hh>
+
 #include <dumux/common/dumuxmessage.hh>
-#include <dumux/common/math.hh>
 #include <dumux/common/parameters.hh>
+#include <dumux/common/partial.hh>
 #include <dumux/common/properties.hh>
+#include <dumux/discretization/method.hh>
+#include <dumux/io/format.hh>
+#include <dumux/io/grid/gridmanager.hh>
+
+#include <dumux/freeflow/navierstokes/velocityoutput.hh>
+#include <dumux/io/vtkoutputmodule.hh>
 
 #include <dumux/linear/istlsolvers.hh>
 #include <dumux/linear/linearalgebratraits.hh>
 #include <dumux/linear/linearsolvertraits.hh>
 
-#include <dumux/nonlinear/newtonsolver.hh>
+#include <dumux/multidomain/newtonsolver.hh>
 
-#include <dumux/assembly/diffmethod.hh>
-#include <dumux/assembly/fvassembler.hh>
-
-#include <dumux/discretization/method.hh>
-
-#include <dumux/io/grid/gridmanager.hh>
-#include <dumux/io/vtkoutputmodule.hh>
-
-#include "pmproblem-reversed.hh"
+#include "ffproperties.hh"
 
 #include "dumux-precice/couplingadapter.hh"
 
-/*!
-  * \brief Returns the pressure at the interface using Darcy's law for reconstruction
-  */
-template<class Problem,
-         class Element,
-         class FVElementGeometry,
-         class ElementVolumeVariables,
-         class SubControlVolumeFace,
-         class ElementFluxVariablesCache>
-auto pressureAtInterface(const Problem &problem,
-                         const Element &element,
-                         const FVElementGeometry &fvGeometry,
-                         const ElementVolumeVariables &elemVolVars,
-                         const SubControlVolumeFace &scvf,
-                         const ElementFluxVariablesCache &elemFluxVarsCache)
+#include <iostream>
+#include <memory>
+#include <type_traits>
+
+//TODO
+// Helper function to put pressure on interface
+
+template<class ElementVolumeVariables, class SubControlVolumeFace>
+auto velocityAtInterface(const ElementVolumeVariables &elemVolVars,
+                         const SubControlVolumeFace &scvf)
 {
-    using Scalar = typename ElementVolumeVariables::VolumeVariables::
-        PrimaryVariables::value_type;
-    const auto &volVars = elemVolVars[scvf.insideScvIdx()];
-
-    const Scalar boundaryFlux = problem.neumann(
-        element, fvGeometry, elemVolVars, elemFluxVarsCache, scvf);
-
-    const auto K = volVars.permeability();
-    const Scalar ccPressure = volVars.pressure();
-    const Scalar mobility = volVars.mobility();
-    const Scalar density = volVars.density();
-
-    // v = -K/mu * (gradP + rho*g)
+    assert(scvf.isFrontal());
+    const double scalarVelocity = elemVolVars[scvf.insideScvIdx()].velocity();
     auto velocity = scvf.unitOuterNormal();
-    velocity *= boundaryFlux;  // TODO check sign
-    velocity /= density;
-
-    // v = -kr/mu*K * (gradP + rho*g) = -mobility*K * (gradP + rho*g)
-    const auto alpha =
-        Dumux::vtmv(scvf.unitOuterNormal(), K, problem.gravity());
-
-    auto distanceVector = scvf.center() - element.geometry().center();
-    distanceVector /= distanceVector.two_norm2();
-    const Scalar ti = Dumux::vtmv(distanceVector, K, scvf.unitOuterNormal());
-
-    return (1 / mobility * (scvf.unitOuterNormal() * velocity) +
-            density * alpha) /
-               ti +
-           ccPressure;
+    velocity[scvf.normalAxis()] = scalarVelocity;
+    return velocity;
 }
 
-template<class Problem, class GridVariables, class SolutionVector>
-void setInterfacePressures(const Problem &problem,
+template<typename TypeTag,
+         class Problem,
+         class Element,
+         class SubControlVolumeFace,
+         class FVElementGeometry,
+         class ElementVolumeVariables,
+         class ElementFluxVariablesCache>
+auto pressureAtInterface(const Problem *problem,
+                         const Element &element,
+                         const SubControlVolumeFace &scvf,
+                         const FVElementGeometry &fvGeometry,
+                         const ElementVolumeVariables &elemVolVars,
+                         const ElementFluxVariablesCache &elemFluxVarsCache)
+{
+    using ElementBoundaryTypes =
+        Dumux::GetPropType<TypeTag, Dumux::Properties::ElementBoundaryTypes>;
+    using LocalResidual =
+        Dumux::GetPropType<TypeTag, Dumux::Properties::LocalResidual>;
+    using NumEqVector = Dumux::NumEqVector<
+        Dumux::GetPropType<TypeTag, Dumux::Properties::PrimaryVariables>>;
+#if DUMUX_VERSION_MAJOR >= 3 && DUMUX_VERSION_MINOR >= 9
+    using SlipVelocityPolicy = Dumux::NavierStokesSlipVelocity<
+        typename FVElementGeometry::GridGeometry::DiscretizationMethod,
+        Dumux::NavierStokes::SlipConditions::BJ>;
+    using FluxHelper = Dumux::NavierStokesMomentumBoundaryFlux<
+        typename FVElementGeometry::GridGeometry::DiscretizationMethod,
+        SlipVelocityPolicy>;
+#else
+    using FluxHelper = Dumux::NavierStokesMomentumBoundaryFluxHelper;
+#endif
+
+    ElementBoundaryTypes elemBcTypes;
+    auto localResidual = LocalResidual(problem);
+    elemBcTypes.update(*problem, element, fvGeometry);
+
+    NumEqVector flux(0.0);
+    NumEqVector neumannFlux(0.0);
+    const auto &scv = fvGeometry.scv(scvf.insideScvIdx());
+    for (const auto &otherScvf : scvfs(fvGeometry, scv)) {
+        if (otherScvf.index() == scvf.index())
+            continue;
+        flux += localResidual.maybeHandleNeumannBoundary(
+            *problem, element, fvGeometry, elemVolVars, elemBcTypes,
+            elemFluxVarsCache, otherScvf);
+    }
+    flux += FluxHelper::fixedPressureMomentumFlux(
+                *problem, fvGeometry, scvf, elemVolVars, elemFluxVarsCache, 0.0,
+                /*zeroNormalVelocityGradient=*/false)[scvf.normalAxis()] *
+            scvf.area();
+    return -1 * scvf.directionSign() * flux / scvf.area();
+}
+
+template<typename MomentumTypeTag,
+         class Problem,
+         class GridVariables,
+         class SolutionVector>
+void setInterfacePressures(const std::shared_ptr<Problem> problem,
                            const GridVariables &gridVars,
                            const SolutionVector &sol,
                            const std::string meshName,
                            const std::string dataName)
 {
-    const auto &gridGeometry = problem.gridGeometry();
-    auto fvGeometry = localView(gridGeometry);
-    auto elemVolVars = localView(gridVars.curGridVolVars());
-    auto elemFluxVarsCache = localView(gridVars.gridFluxVarsCache());
-
-    auto &couplingParticipant = Dumux::Precice::CouplingAdapter::getInstance();
-
-    for (const auto &element : elements(gridGeometry.gridView())) {
-        fvGeometry.bindElement(element);
-        elemVolVars.bindElement(element, fvGeometry, sol);
-
-        //sstd::cout << "Pressure by reconstruction" << std::endl;
-        for (const auto &scvf : scvfs(fvGeometry)) {
-            if (couplingParticipant.isCoupledEntity(scvf.index())) {
-                //TODO: What to do here?
-                const double p =
-                    pressureAtInterface(problem, element, gridGeometry,
-                                        elemVolVars, scvf, elemFluxVarsCache);
-                couplingParticipant.writeScalarQuantityOnFace(
-                    meshName, dataName, scvf.index(), p);
-            }
-        }
-    }
-}
-
-/*!
-  * \brief Returns the velocity at the interface using Darcy's law for reconstruction
-  */
-template<class FluxVariables,
-         class Problem,
-         class Element,
-         class FVElementGeometry,
-         class ElementVolumeVariables,
-         class SubControlVolumeFace,
-         class ElementFluxVariablesCache>
-auto velocityAtInterface(const Problem &problem,
-                         const Element &element,
-                         const FVElementGeometry &fvGeometry,
-                         const ElementVolumeVariables &elemVolVars,
-                         const SubControlVolumeFace &scvf,
-                         const ElementFluxVariablesCache &elemFluxVarsCache)
-{
-    const int phaseIdx = 0;
-    FluxVariables fluxVars;
-    fluxVars.init(problem, element, fvGeometry, elemVolVars, scvf,
-                  elemFluxVarsCache);
-    auto upwindTerm = [phaseIdx](const auto &volVars) {
-        return volVars.mobility(phaseIdx);
-    };
-    const auto scalarVelocity =
-        fluxVars.advectiveFlux(phaseIdx, upwindTerm) / scvf.area();
-    return scalarVelocity;
-}
-
-template<class FluxVariables,
-         class Problem,
-         class GridVariables,
-         class SolutionVector>
-void setInterfaceVelocities(const Problem &problem,
-                            const GridVariables &gridVars,
-                            const SolutionVector &sol,
-                            const std::string meshName,
-                            const std::string dataName)
-{
-    const auto &gridGeometry = problem.gridGeometry();
+    const auto &gridGeometry = problem->gridGeometry();
     auto fvGeometry = localView(gridGeometry);
     auto elemVolVars = localView(gridVars.curGridVolVars());
     auto elemFluxVarsCache = localView(gridVars.gridFluxVarsCache());
@@ -193,9 +150,38 @@ void setInterfaceVelocities(const Problem &problem,
         for (const auto &scvf : scvfs(fvGeometry)) {
             if (couplingParticipant.isCoupledEntity(scvf.index())) {
                 //TODO: What to do here?
-                const double v = velocityAtInterface<FluxVariables>(
-                    problem, element, fvGeometry, elemVolVars, scvf,
+                const auto p = pressureAtInterface<MomentumTypeTag>(
+                    problem.get(), element, scvf, fvGeometry, elemVolVars,
                     elemFluxVarsCache);
+                couplingParticipant.writeScalarQuantityOnFace(
+                    meshName, dataName, scvf.index(), p);
+            }
+        }
+    }
+}
+
+template<class MomentumProblem, class GridVariables, class SolutionVector>
+void setInterfaceVelocities(const MomentumProblem &problem,
+                            const GridVariables &gridVars,
+                            const SolutionVector &sol,
+                            const std::string meshName,
+                            const std::string dataName)
+{
+    const auto &gridGeometry = problem.gridGeometry();
+    auto fvGeometry = localView(gridGeometry);
+    auto elemVolVars = localView(gridVars.curGridVolVars());
+
+    auto &couplingParticipant = Dumux::Precice::CouplingAdapter::getInstance();
+
+    for (const auto &element : elements(gridGeometry.gridView())) {
+        fvGeometry.bindElement(element);
+        elemVolVars.bindElement(element, fvGeometry, sol);
+
+        for (const auto &scvf : scvfs(fvGeometry)) {
+            if (couplingParticipant.isCoupledEntity(scvf.index())) {
+                //TODO: What to do here?
+                const auto v =
+                    velocityAtInterface(elemVolVars, scvf)[scvf.normalAxis()];
                 couplingParticipant.writeScalarQuantityOnFace(
                     meshName, dataName, scvf.index(), v);
             }
@@ -203,21 +189,17 @@ void setInterfaceVelocities(const Problem &problem,
     }
 }
 
-template<class FluxVariables,
-         class Problem,
-         class GridVariables,
-         class SolutionVector>
+template<class MomentumProblem, class GridVariables, class SolutionVector>
 std::tuple<double, double, double> writeVelocitiesOnInterfaceToFile(
     const std::string &meshName,
     const std::string &filename,
-    const Problem &problem,
+    const std::shared_ptr<MomentumProblem> problem,
     const GridVariables &gridVars,
     const SolutionVector &sol)
 {
-    const auto &gridGeometry = problem.gridGeometry();
+    const auto &gridGeometry = problem->gridGeometry();
     auto fvGeometry = localView(gridGeometry);
     auto elemVolVars = localView(gridVars.curGridVolVars());
-    auto elemFluxVarsCache = localView(gridVars.gridFluxVarsCache());
 
     const auto &couplingParticipant =
         Dumux::Precice::CouplingAdapter::getInstance();
@@ -236,7 +218,6 @@ std::tuple<double, double, double> writeVelocitiesOnInterfaceToFile(
     for (const auto &element : elements(gridGeometry.gridView())) {
         fvGeometry.bind(element);
         elemVolVars.bind(element, fvGeometry, sol);
-        elemFluxVarsCache.bind(element, fvGeometry, elemVolVars);
 
         for (const auto &scvf : scvfs(fvGeometry)) {
             if (couplingParticipant.isCoupledEntity(scvf.index())) {
@@ -245,9 +226,8 @@ std::tuple<double, double, double> writeVelocitiesOnInterfaceToFile(
                      i < couplingParticipant.getMeshDimensions(meshName); ++i) {
                     ofs << pos[i] << ",";
                 }
-                const double v = velocityAtInterface<FluxVariables>(
-                    problem, element, fvGeometry, elemVolVars, scvf,
-                    elemFluxVarsCache);
+                const double v =
+                    velocityAtInterface(elemVolVars, scvf)[scvf.normalAxis()];
                 max = std::max(v, max);
                 min = std::min(v, min);
                 sum += v;
@@ -261,17 +241,21 @@ std::tuple<double, double, double> writeVelocitiesOnInterfaceToFile(
     }
 
     ofs.close();
+
     return std::make_tuple(min, max, sum);
 }
 
-template<class Problem, class GridVariables, class SolutionVector>
+template<typename MomentumTypeTag,
+         class Problem,
+         class GridVariables,
+         class SolutionVector>
 void writePressuresOnInterfaceToFile(const std::string &meshName,
                                      const std::string &filename,
-                                     const Problem &problem,
+                                     const std::shared_ptr<Problem> problem,
                                      const GridVariables &gridVars,
                                      const SolutionVector &sol)
 {
-    const auto &gridGeometry = problem.gridGeometry();
+    const auto &gridGeometry = problem->gridGeometry();
     auto fvGeometry = localView(gridGeometry);
     auto elemVolVars = localView(gridVars.curGridVolVars());
     auto elemFluxVarsCache = localView(gridVars.gridFluxVarsCache());
@@ -298,9 +282,9 @@ void writePressuresOnInterfaceToFile(const std::string &meshName,
                      i < couplingParticipant.getMeshDimensions(meshName); ++i) {
                     ofs << pos[i] << ",";
                 }
-                const double p =
-                    pressureAtInterface(problem, element, gridGeometry,
-                                        elemVolVars, scvf, elemFluxVarsCache);
+                const double p = pressureAtInterface<MomentumTypeTag>(
+                    problem.get(), element, scvf, fvGeometry, elemVolVars,
+                    elemFluxVarsCache);
                 ofs << p << "\n";
             }
         }
@@ -323,34 +307,53 @@ try {
     // parse command line arguments and input file
     Parameters::init(argc, argv);
 
-    using DarcyTypeTag = Properties::TTag::DarcyOneP;
+    // Define the sub problem type tags
+    using MomentumTypeTag = Properties::TTag::FreeFlowSubMomentum;
+    using MassTypeTag = Properties::TTag::FreeFlowSubMass;
 
-    using DarcyGridManager =
-        Dumux::GridManager<GetPropType<DarcyTypeTag, Properties::Grid>>;
-    DarcyGridManager darcyGridManager;
-    darcyGridManager.init("Darcy");  // pass parameter group
+    // try to create a grid (from the given grid file or the input file)
+    using FreeFlowGridManager =
+        Dumux::GridManager<GetPropType<MomentumTypeTag, Properties::Grid>>;
+    FreeFlowGridManager freeFlowGridManager;
+    freeFlowGridManager.init("FreeFlow");  // pass parameter group
 
     // we compute on the leaf grid view
-    const auto &darcyGridView = darcyGridManager.grid().leafGridView();
+    const auto &freeFlowGridView = freeFlowGridManager.grid().leafGridView();
 
     // create the finite volume grid geometry
-    using DarcyGridGeometry =
-        GetPropType<DarcyTypeTag, Properties::GridGeometry>;
-    auto darcyGridGeometry = std::make_shared<DarcyGridGeometry>(darcyGridView);
-    darcyGridGeometry->update(darcyGridManager.grid().leafGridView());
+    using MomentumGridGeometry =
+        GetPropType<MomentumTypeTag, Properties::GridGeometry>;
+    auto momentumGridGeometry =
+        std::make_shared<MomentumGridGeometry>(freeFlowGridView);
+    using MassGridGeometry = GetPropType<MassTypeTag, Properties::GridGeometry>;
+    auto massGridGeometry =
+        std::make_shared<MassGridGeometry>(freeFlowGridView);
 
-    using DarcyProblem = GetPropType<DarcyTypeTag, Properties::Problem>;
-    auto darcyProblem = std::make_shared<DarcyProblem>(darcyGridGeometry);
+    // create the coupling manager to couple the two subproblems of the freeflow participant
+    using CouplingManager =
+        GetPropType<MomentumTypeTag, Properties::CouplingManager>;
+    auto couplingManager = std::make_shared<CouplingManager>();
+    constexpr auto momentumIdx = CouplingManager::freeFlowMomentumIndex;
+    constexpr auto massIdx = CouplingManager::freeFlowMassIndex;
+
+    // the problem (initial and boundary conditions)
+    using MomentumProblem = GetPropType<MomentumTypeTag, Properties::Problem>;
+    auto momentumProblem = std::make_shared<MomentumProblem>(
+        momentumGridGeometry, couplingManager);
+    using MassProblem = GetPropType<MassTypeTag, Properties::Problem>;
+    auto massProblem =
+        std::make_shared<MassProblem>(massGridGeometry, couplingManager);
 
     // the solution vector
-    GetPropType<DarcyTypeTag, Properties::SolutionVector> sol;
-    sol.resize(darcyGridGeometry->numDofs());
+    using Traits = MultiDomainTraits<MomentumTypeTag, MassTypeTag>;
+    using SolutionVector = Traits::SolutionVector;
+    SolutionVector sol;
 
     // Initialize preCICE.Tell preCICE about:
     // - Name of solver
     // - What rank of how many ranks this instance is
     // Configure preCICE. For now the config file is hardcoded.
-    //couplingParticipant.createInstance( "darcy", mpiHelper.rank(), mpiHelper.size() );
+    //couplingParticipant.createInstance( "FreeFlow", mpiHelper.rank(), mpiHelper.size() );
     std::string preciceConfigFilename = "precice-config.xml";
     //    if (argc == 3)
     //      preciceConfigFilename = argv[2];
@@ -358,14 +361,14 @@ try {
         preciceConfigFilename = argv[argc - 1];
 
     auto &couplingParticipant = Dumux::Precice::CouplingAdapter::getInstance();
-    couplingParticipant.announceSolver("Darcy", preciceConfigFilename,
+    couplingParticipant.announceSolver("FreeFlow", preciceConfigFilename,
                                        mpiHelper.rank(), mpiHelper.size());
 
-    const std::string meshName("DarcyMesh");
+    const std::string meshName("FreeFlowMesh");  // mesh name
     const int dim = couplingParticipant.getMeshDimensions(meshName);
-    std::cout << dim << "  " << int(DarcyGridGeometry::GridView::dimension)
+    std::cout << dim << "  " << int(MassGridGeometry::GridView::dimension)
               << std::endl;
-    if (dim != int(DarcyGridGeometry::GridView::dimension))
+    if (dim != int(MassGridGeometry::GridView::dimension))
         DUNE_THROW(Dune::InvalidStateException, "Dimensions do not match");
 
     // GET mesh corodinates
@@ -376,14 +379,16 @@ try {
     std::vector<double> coords;  //( dim * vertexSize );
     std::vector<int> coupledScvfIndices;
 
-    for (const auto &element : elements(darcyGridView)) {
-        auto fvGeometry = localView(*darcyGridGeometry);
+    for (const auto &element : elements(freeFlowGridView)) {
+        auto fvGeometry = localView(*momentumGridGeometry);
         fvGeometry.bindElement(element);
 
         for (const auto &scvf : scvfs(fvGeometry)) {
+            if (!scvf.isFrontal())
+                continue;
             static constexpr auto eps = 1e-7;
             const auto &pos = scvf.center();
-            if (pos[1] > darcyGridGeometry->bBoxMax()[1] - eps) {
+            if (pos[1] < momentumGridGeometry->bBoxMin()[1] + eps) {
                 if (pos[0] > xMin - eps && pos[0] < xMax + eps) {
                     coupledScvfIndices.push_back(scvf.index());
                     for (const auto p : pos)
@@ -398,46 +403,57 @@ try {
 
     const std::string dataNameV("Velocity");
     const std::string dataNameP("Pressure");
-    couplingParticipant.announceQuantity(meshName, dataNameP);
     couplingParticipant.announceQuantity(meshName, dataNameV);
+    couplingParticipant.announceQuantity(meshName, dataNameP);
 
-    darcyProblem->applyInitialSolution(sol);
+    // apply initial solution for instationary problems
+    momentumProblem->applyInitialSolution(sol[momentumIdx]);
+    massProblem->applyInitialSolution(sol[massIdx]);
 
     // the grid variables
-    using DarcyGridVariables =
-        GetPropType<DarcyTypeTag, Properties::GridVariables>;
-    auto darcyGridVariables =
-        std::make_shared<DarcyGridVariables>(darcyProblem, darcyGridGeometry);
-    darcyGridVariables->init(sol);
+    using MomentumGridVariables =
+        GetPropType<MomentumTypeTag, Properties::GridVariables>;
+    auto momentumGridVariables = std::make_shared<MomentumGridVariables>(
+        momentumProblem, momentumGridGeometry);
+    using MassGridVariables =
+        GetPropType<MassTypeTag, Properties::GridVariables>;
+    auto massGridVariables =
+        std::make_shared<MassGridVariables>(massProblem, massGridGeometry);
+
+    // initialize the coupling manager and the grid variables of the subproblems
+    couplingManager->init(
+        momentumProblem, massProblem,
+        std::make_tuple(momentumGridVariables, massGridVariables), sol);
+    momentumGridVariables->init(sol[momentumIdx]);
+    massGridVariables->init(sol[massIdx]);
 
     // intialize the vtk output module
-    const auto darcyName =
-        getParam<std::string>("Problem.Name") + "_" + darcyProblem->name();
+    using IOFields = GetPropType<MassTypeTag, Properties::IOFields>;
+    VtkOutputModule freeFlowVtkWriter(*massGridVariables, sol[massIdx],
+                                      massProblem->name());
+    IOFields::initOutputModule(freeFlowVtkWriter);
+    freeFlowVtkWriter.addVelocityOutput(
+        std::make_shared<NavierStokesVelocityOutput<MassGridVariables>>());
+    freeFlowVtkWriter.addField(massProblem->getAnalyticalVelocityX(),
+                               "analyticalV_x");
+    freeFlowVtkWriter.write(0.0);
 
-    VtkOutputModule<DarcyGridVariables,
-                    GetPropType<DarcyTypeTag, Properties::SolutionVector>>
-        darcyVtkWriter(*darcyGridVariables, sol, darcyProblem->name());
-    using DarcyVelocityOutput =
-        GetPropType<DarcyTypeTag, Properties::VelocityOutput>;
-    darcyVtkWriter.addVelocityOutput(
-        std::make_shared<DarcyVelocityOutput>(*darcyGridVariables));
-    GetPropType<DarcyTypeTag, Properties::IOFields>::initOutputModule(
-        darcyVtkWriter);
-    darcyVtkWriter.write(0.0);
-
-    using FluxVariables = GetPropType<DarcyTypeTag, Properties::FluxVariables>;
     if (couplingParticipant.requiresToWriteInitialData()) {
-        //TODO
-        setInterfaceVelocities<FluxVariables>(
-            *darcyProblem, *darcyGridVariables, sol, meshName, dataNameV);
-        couplingParticipant.writeQuantityToOtherSolver(meshName, dataNameV);
+        setInterfacePressures<MomentumTypeTag>(
+            momentumProblem, *momentumGridVariables, sol[momentumIdx], meshName,
+            dataNameP);
+        couplingParticipant.writeQuantityToOtherSolver(meshName, dataNameP);
     }
     couplingParticipant.initialize();
 
     // the assembler for a stationary problem
-    using Assembler = FVAssembler<DarcyTypeTag, DiffMethod::numeric>;
+    using Assembler =
+        MultiDomainFVAssembler<Traits, CouplingManager, DiffMethod::numeric>;
     auto assembler = std::make_shared<Assembler>(
-        darcyProblem, darcyGridGeometry, darcyGridVariables);
+        std::make_tuple(momentumProblem, massProblem),
+        std::make_tuple(momentumGridGeometry, massGridGeometry),
+        std::make_tuple(momentumGridVariables, massGridVariables),
+        couplingManager);
 
     // the linear solver
     using LinearSolver =
@@ -446,8 +462,9 @@ try {
     auto linearSolver = std::make_shared<LinearSolver>();
 
     // the non-linear solver
-    using NewtonSolver = Dumux::NewtonSolver<Assembler, LinearSolver>;
-    NewtonSolver nonLinearSolver(assembler, linearSolver);
+    using NewtonSolver =
+        MultiDomainNewtonSolver<Assembler, LinearSolver, CouplingManager>;
+    NewtonSolver nonLinearSolver(assembler, linearSolver, couplingManager);
 
     double preciceDt = couplingParticipant.getMaxTimeStepSize();
     auto dt = preciceDt;
@@ -462,15 +479,25 @@ try {
             sol_checkpoint = sol;
         }
 
-        couplingParticipant.readQuantityFromOtherSolver(meshName, dataNameP,
+        couplingParticipant.readQuantityFromOtherSolver(meshName, dataNameV,
                                                         dt);
-
         // solve the non-linear system
         nonLinearSolver.solve(sol);
-        setInterfaceVelocities<FluxVariables>(
-            *darcyProblem, *darcyGridVariables, sol, meshName, dataNameV);
-        couplingParticipant.writeQuantityToOtherSolver(meshName, dataNameV);
 
+        writeVelocitiesOnInterfaceToFile(
+            meshName, Dumux::Fmt::format("ff_interface_velocities_{}", vtkTime),
+            momentumProblem, *momentumGridVariables, sol[momentumIdx]);
+        writePressuresOnInterfaceToFile<MomentumTypeTag>(
+            meshName, Dumux::Fmt::format("ff_interface_pressures_{}", vtkTime),
+            momentumProblem, *momentumGridVariables, sol[momentumIdx]);
+
+        setInterfacePressures<MomentumTypeTag>(
+            momentumProblem, *momentumGridVariables, sol[momentumIdx], meshName,
+            dataNameP);
+        couplingParticipant.writeQuantityToOtherSolver(meshName, dataNameP);
+        //Read checkpoint
+        freeFlowVtkWriter.write(vtkTime);
+        vtkTime += 1.;
         couplingParticipant.advance(dt);
         preciceDt = couplingParticipant.getMaxTimeStepSize();
         dt = std::min(preciceDt, dt);
@@ -478,26 +505,22 @@ try {
         ++iter;
 
         if (couplingParticipant.requiresToReadCheckpoint()) {
-            //Read checkpoint
-            darcyVtkWriter.write(vtkTime);
-            vtkTime += 1.;
             sol = sol_checkpoint;
-            darcyGridVariables->update(sol);
-            darcyGridVariables->advanceTimeStep();
+            momentumGridVariables->update(sol[momentumIdx]);
+            massGridVariables->update(sol[massIdx]);
+            momentumGridVariables->advanceTimeStep();
+            massGridVariables->advanceTimeStep();
         } else  // coupling successful
         {
             // write vtk output
-            darcyVtkWriter.write(vtkTime);
+            freeFlowVtkWriter.write(vtkTime);
         }
     }
-    // write vtk output
-    darcyVtkWriter.write(1.0);
-
-    couplingParticipant.finalize();
-
     ////////////////////////////////////////////////////////////
     // finalize, print dumux message to say goodbye
     ////////////////////////////////////////////////////////////
+
+    couplingParticipant.finalize();
 
     // print dumux end message
     if (mpiHelper.rank() == 0) {